% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_objects_minimal.R
\name{analyze_objects_minimal}
\alias{analyze_objects_minimal}
\alias{plot.anal_obj_minimal}
\alias{plot.anal_obj_ls_minimal}
\title{Analyzes objects in an image}
\usage{
analyze_objects_minimal(
  img,
  segment_objects = TRUE,
  reference = FALSE,
  reference_area = NULL,
  back_fore_index = "R/(G/B)",
  fore_ref_index = "B-R",
  reference_larger = FALSE,
  reference_smaller = FALSE,
  pattern = NULL,
  parallel = FALSE,
  workers = NULL,
  watershed = TRUE,
  fill_hull = FALSE,
  opening = FALSE,
  closing = FALSE,
  filter = FALSE,
  erode = FALSE,
  dilate = FALSE,
  invert = FALSE,
  object_size = "medium",
  index = "NB",
  r = 1,
  g = 2,
  b = 3,
  re = 4,
  nir = 5,
  threshold = "Otsu",
  tolerance = NULL,
  extension = NULL,
  lower_noise = 0.1,
  lower_size = NULL,
  upper_size = NULL,
  topn_lower = NULL,
  topn_upper = NULL,
  lower_eccent = NULL,
  upper_eccent = NULL,
  lower_circ = NULL,
  upper_circ = NULL,
  plot = TRUE,
  show_original = TRUE,
  show_contour = TRUE,
  contour_col = "red",
  contour_size = 1,
  col_foreground = NULL,
  col_background = NULL,
  marker = FALSE,
  marker_col = NULL,
  marker_size = NULL,
  save_image = FALSE,
  max_pixels = 4e+06,
  prefix = "proc_",
  dir_original = NULL,
  dir_processed = NULL,
  verbose = TRUE
)

\method{plot}{anal_obj_minimal}(
  x,
  which = "measure",
  measure = "area",
  type = c("density", "histogram"),
  ...
)

\method{plot}{anal_obj_ls_minimal}(
  x,
  which = "measure",
  measure = "area",
  type = c("density", "histogram"),
  ...
)
}
\arguments{
\item{img}{The image to be analyzed.}

\item{segment_objects}{Segment objects in the image? Defaults to \code{TRUE}. If
\code{segment_objects = FALSE}, the objects are not segmented and the entire image
is analyzed (useful for whole-canopy indices).}

\item{reference}{Logical to indicate if a reference object is present in the image.}

\item{reference_area}{The known area of the reference object.}

\item{back_fore_index}{A character value to indicate the index to segment the
foreground (objects and reference) from the background. Defaults to \code{"R/(G/B)"}.}

\item{fore_ref_index}{A character value to indicate the index to segment
objects and the reference object. Defaults to \code{"B-R"}.}

\item{reference_larger, reference_smaller}{Logical argument indicating when the
larger/smaller object in the image must be used as the reference object.}

\item{pattern}{A pattern of file name used to identify images to be imported.}

\item{parallel}{If \code{TRUE} processes the images asynchronously in parallel.}

\item{workers}{Number of parallel processes.}

\item{watershed}{If \code{TRUE} (default) performs watershed-based object detection.}

\item{fill_hull}{Fill holes in the binary image? Defaults to \code{FALSE}.}

\item{opening, closing, filter, erode, dilate}{\strong{Morphological operations (brush
size)}, applied in hierarchical order.}

\item{invert}{Inverts the binary image if desired.}

\item{object_size}{The size of the object (\code{"small"}, \code{"medium"}, \code{"large"}, \code{"elarge"}).
Used to automatically set up \code{tolerance} and \code{extension} parameters.}

\item{index}{A character value specifying the index for conversion to binary image.}

\item{r, g, b, re, nir}{The red, green, blue, red-edge, and near-infrared bands, respectively.}

\item{threshold}{The threshold method (\code{"Otsu"}, \code{"Adaptive"}, or numeric).}

\item{tolerance}{Minimum height for watershed segmentation.}

\item{extension}{Neighborhood radius for watershed segmentation.}

\item{lower_noise}{Objects with area less than \code{lower_noise} * mean area are removed.}

\item{lower_size, upper_size}{Lower and upper size limits for object filtering.}

\item{topn_lower, topn_upper}{Select top \code{n} objects based on area.}

\item{lower_eccent, upper_eccent, lower_circ, upper_circ}{Eccentricity and circularity limits.}

\item{plot}{Show image after processing?}

\item{show_original}{Show count objects on the original image?}

\item{show_contour}{Show contour line around objects?}

\item{contour_col, contour_size}{Color and size for contour line.}

\item{col_foreground, col_background}{Foreground and background colors.}

\item{marker, marker_col, marker_size}{Type, color, and size of object marker.}

\item{save_image}{Save the image after processing?}

\item{max_pixels}{Maximum number of pixels to render when plotting or saving.}

\item{prefix}{Prefix for processed images.}

\item{dir_original, dir_processed}{Directories for original and processed images.}

\item{verbose}{If \code{TRUE} (default) shows progress summary in console.}

\item{x}{An object of class \code{anal_obj}.}

\item{which}{Which to plot. Either 'measure' (object measures) or 'index'
(object index). Defaults to \code{"measure"}.}

\item{measure}{The measure to plot. Defaults to \code{"area"}.}

\item{type}{The type of plot. Either \code{"hist"} or \code{"density"}. Partial matches
are recognized.}

\item{...}{Additional arguments passed to methods.}
}
\description{
A lighter option to \code{\link[=analyze_objects]{analyze_objects()}}
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("soybean_touch.jpg")
obj <- analyze_objects(img)
obj$statistics

}

if (interactive()) {
library(pliman)

img <- image_pliman("soy_green.jpg")
# Segment the foreground (grains) using the normalized blue index (NB, default)
# Shows the average value of the blue index in each object

rgb <- analyze_objects_minimal(img)
# density of area
plot(rgb)

# histogram of area
plot(rgb, type = "histogram") # or 'hist'
}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
