% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{as.numeric.image}
\alias{as.numeric.image}
\title{Coerce \code{image} pixel values to numeric vector}
\usage{
\method{as.numeric}{image}(x, raw_to_scale = TRUE, ...)
}
\arguments{
\item{x}{An \code{image} object.}

\item{raw_to_scale}{Logical; if \code{TRUE} and image is \code{raw}, scales 0..255 to 0..1. Default \code{TRUE}.}

\item{...}{Unused.}
}
\description{
Coerce \code{image} pixel values to numeric vector
}
