% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{as.raster.image}
\alias{as.raster.image}
\title{Convert \code{image} to standard R raster matrix}
\usage{
\method{as.raster}{image}(x, ...)
}
\arguments{
\item{x}{An \code{image} object.}

\item{...}{Unused.}
}
\description{
Convert \code{image} to standard R raster matrix
}
