% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{as_image}
\alias{as_image}
\title{Create an \code{Image} object}
\usage{
as_image(data, ...)
}
\arguments{
\item{data}{A vector or array containing the pixel intensities of an image.
If missing, the default 1x1 zero-filled array is used.}

\item{...}{Additional arguments passed to \code{\link[EBImage:Image]{EBImage::Image()}}.}
}
\value{
An \code{Image} object.
}
\description{
This function is a simple wrapper around \code{\link[EBImage:Image]{EBImage::Image()}}.
}
\examples{
img <-
as_image(rnorm(150 * 150 * 3),
         dim = c(150, 150, 3),
         colormode = 'Color')
plot(img)
}
