% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_yolo_autolabel.R
\name{c.yolo_fewshot_model}
\alias{c.yolo_fewshot_model}
\title{Combine Few-Shot Models}
\usage{
\method{c}{yolo_fewshot_model}(...)
}
\arguments{
\item{...}{\code{yolo_fewshot_model} objects to combine.}
}
\value{
A \code{yolo_fewshot_multimodel} object.
}
\description{
Combines multiple \code{yolo_fewshot_model} objects into a multi-class few-shot model.
}
