% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morphological.R
\name{contour}
\alias{contour}
\title{Extract object contours from a labeled image}
\usage{
contour(labels)
}
\arguments{
\item{labels}{A labeled \code{image} object or a matrix containing integer labels. Typically
the result of \code{\link[=image_watershed]{image_watershed()}} or similar segmentation.}
}
\value{
A list where each element is a matrix of (x, y) coordinates for the
contour of the respective object.
}
\description{
Extracts the external contour of each object in a labeled matrix (e.g., from
watershed segmentation). It returns a list of coordinates for each object.
}
\examples{
library(pliman)
# generate a binary image
img <- image_pliman("soybean_touch.jpg")
bin <- image_binary(img, "B")[[1]]
# watershed segmentation
wat <- image_watershed(bin)
# extract contours
cont <- contour(wat)
plot(img)
plot_contour(cont, col = "red", lwd = 2)
}
