% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{contours}
\alias{contours}
\title{Contour outlines from five leaves}
\format{
A list with five objects
\itemize{
\item \code{leaf_1}
\item \code{leaf_2}
\item \code{leaf_3}
\item \code{leaf_4}
\item \code{leaf_5}
}

Each object is a \code{data.frame} with the coordinates for the outline perimeter
}
\source{
Personal data. The images were obtained in the Flavia data set
downlodable at \url{https://flavia.sourceforge.net/}
}
\description{
A list of contour outlines from five leaves. It may be used as
example in some functions such as \code{\link[=efourier]{efourier()}}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{data}
