% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_extract_raster}
\alias{cpp_extract_raster}
\title{Fast C++ raster extraction engine without GDAL}
\usage{
cpp_extract_raster(
  values,
  n_row,
  n_col,
  n_lyr,
  bbox_raster,
  geoms_r,
  fun = as.character(c("none")),
  exact = FALSE,
  return_coverage_area = FALSE,
  subdiv = 10L,
  summarize_quantiles = as.numeric(c(0.05, 0.975))
)
}
\arguments{
\item{values}{Matrix of raster values (ncell x nlyr).}

\item{n_row}{Number of raster rows.}

\item{n_col}{Number of raster columns.}

\item{n_lyr}{Number of raster layers.}

\item{bbox_raster}{Vector c(xmin, xmax, ymin, ymax).}

\item{geoms_r}{Rcpp::List of polygon geometries for each feature.}

\item{fun}{CharacterVector of summary functions ("none", "mean", "median", "sum", "min", "max", "sd", "count", "quantile", "quantiles").}

\item{exact}{Logical, whether to compute exact pixel area fraction coverage.}

\item{return_coverage_area}{Logical, whether to include coverage_area column in raw pixel output.}

\item{subdiv}{Number of subdivisions per dimension when exact = TRUE.}

\item{summarize_quantiles}{NumericVector of quantile probabilities (between 0 and 1) when "quantiles" is requested.}
}
\description{
Fast C++ raster extraction engine without GDAL
}
