% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erosao_conservativa}
\alias{erosao_conservativa}
\title{Erosão conservativa rápida (pipeline completo)}
\usage{
erosao_conservativa(img, raio_erosao = 3L, raio_dilatacao = 3L, forma = "disc")
}
\description{
Erosão conservativa rápida (pipeline completo)
}
