% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_free_ram_cpp}
\alias{get_free_ram_cpp}
\title{Fast C++ retrieval of available system physical RAM in bytes (Windows, Linux, macOS)}
\usage{
get_free_ram_cpp()
}
\description{
Fast C++ retrieval of available system physical RAM in bytes (Windows, Linux, macOS)
}
