% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Get object measures}
\usage{
get_measures(object, dpi = NULL, id = NULL, measure = NULL, true_value = NULL)
}
\arguments{
\item{object}{An object computed with \code{\link[=count_objects]{count_objects()}} or \code{\link[=leaf_area]{leaf_area()}}.}

\item{dpi}{A known resolution of the image in DPI (dots per inch).}

\item{id}{An object in the image to indicate a known value.}

\item{measure}{The measure of the known object. One of the following:
\itemize{
\item \code{"area"} The known area of the object.
\item \code{"perimeter"} The known perimeter of the object.
\item \code{"radius_mean"} The known radius of the object.
\item \code{"radius_min"} The known minimum radius of the object. If the object is a
square, then the \code{radius_min} of such object will be \code{L/2} where \code{L} is the
length of the square side.
\item \code{"radius_max"} The known maximum radius of the object. If the object is a
square, then the \code{radius_max} of such object according to the Pythagorean
theorem will be \verb{L x sqrt(2) / 2} where \code{L} is the length of the square side.
}}

\item{true_value}{The know value of the object \code{id.}}
}
\value{
A data frame with the object \code{id} and the measures. If any of the
\code{"area"}, \code{"perimeter"}, \code{"radius_mean"}, \code{"radius_min"} or \code{"radius_max"}
is informed, the informed measure will be returned in the unit given in
\code{true_value} and the other measures in pixels. If \code{dpi} is informed, then
all the measures will be adjusted to the known \code{dpi} and the values
returned in cm2 (area) and cm (the other measures).
}
\description{
Computes object measures (area, perimeter, radius) by using either a known
resolution (dpi) or an object with known measurements.
}
\examples{
img <- img <- image_import(system.file("tmp_images/objects_150dpi.png", package = "pliman"))
image_show(img)
# Image with four objects with a known resolution of 150 dpi
# Higher square: 10 x 10 cm
# Lower square: 5 x 5 cm
# Rectangle: 4 x 2 cm
# Circle: 3 cm in diameter

# Count the objects using the blue band to segment the image
results <-
   count_objects(img,
                 marker = "text",
                 index = "B")

# Get object measures by declaring the known resolution in dots per inch
(measures <- get_measures(results, dpi = 150))

# Calculated diagonal of the object 1
# 10 * sqrt(2) = 14.14

# Observed diagonal of the object 1
measures[1, "radius_max"] * 2


# Get object measures by declaring the known area.
get_measures(results,
             id = 1,
             measure = "area",
             true_value = 100)
}
