% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_pliman_viewer}
\alias{get_pliman_viewer}
\title{Get the value of the pliman_viewer option}
\usage{
get_pliman_viewer()
}
\value{
The current value of the pliman_viewer option.
}
\description{
Retrieves the current value of the pliman_viewer option used in the package.
}
