% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ggplot_color}
\alias{ggplot_color}
\title{ggplot2-like colors generation}
\usage{
ggplot_color(n = 1)
}
\arguments{
\item{n}{The number of colors. This works well for up to about eight colours,
but after that it becomes hard to tell the different colours apart.}
}
\description{
Generate ggplot2
}
\examples{

library(pliman)
ggplot_color(n = 3)
}
