% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_segmentation.R
\name{image_color_labels}
\alias{image_color_labels}
\title{Colorize Labeled Images}
\usage{
image_color_labels(labels)
}
\arguments{
\item{labels}{A labeled image (an \code{image} object or a matrix containing integer labels).}
}
\value{
A color \code{image} object.
}
\description{
Colorizes a labeled image by allocating a different color to each object.
Background pixels (value 0) are colorized with black.
}
\examples{
\dontrun{
  library(pliman)
  bin <- image_pliman("soybean_touch.jpg") |> image_binary(index = "B")
  lbl <- image_bwlabel(bin$B)
  color_lbl <- image_color_labels(lbl)
  plot(color_lbl)
}
}
