% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_correction_pick}
\alias{image_correction_pick}
\title{Interactive Color Correction}
\usage{
image_correction_pick(
  img,
  known_colors = NULL,
  color_chips = 24,
  model = c("ccm", "affine", "white_balance", "cubic", "root_polynomial"),
  lambda = 1e-04
)
}
\arguments{
\item{img}{An \code{image} object to be corrected.}

\item{known_colors}{A \code{data.frame} containing target reference values with \verb{R, G, B}
columns. If \code{NULL} and \code{color_chips == 24}, defaults to \code{colorchecker_ref()}.}

\item{color_chips}{The number of color patches to sample interactively (default \code{24}).}

\item{model}{The calibration model to use (\code{"ccm"}, \code{"affine"}, \code{"white_balance"},
\code{"cubic"}, or \code{"root_polynomial"}). Default \code{"ccm"}.}

\item{lambda}{Ridge regularization factor (default \code{1e-4}).}
}
\value{
An \code{image} object with corrected colors.
}
\description{
Interactively calibrates the color of an image by prompting the user to
sample color patches. It calls \code{pliman::pick_rgb_area()} to pause and
prompt the user to click on \code{color_chips} patches in the image. These
sampled colors are then mapped to \code{known_colors} using the specified \code{model}.
}
\examples{
if(interactive()){
library(pliman)

# Sample 4 known chips (e.g. blue, red, yellow, white)
known_colors <- data.frame(
 id = c(1, 2, 3, 4),
 R = c(25, 186, 245, 249),
 G = c(55, 26, 205, 242),
 B = c(135, 51, 0, 238)
)
img <- image_pliman("colorcheck.jpg")
img_cor <- image_correction_pick(img, known_colors, color_chips = 4)
image_combine(img, img_cor)
}
}
