% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_create}
\alias{image_create}
\title{Create an \code{image} object of a given color}
\usage{
image_create(color, width = 200, heigth = 200, plot = FALSE)
}
\arguments{
\item{color}{either a color name (as listed by \code{\link[grDevices:colors]{grDevices::colors()}}), or a hexadecimal
string of the form \code{"#rrggbb"}.}

\item{width, heigth}{The width and heigth of the image in pixel units.}

\item{plot}{Plots the image after creating it? Defaults to \code{FALSE}.}
}
\value{
An object of class \code{image}.
}
\description{
image_create() can be used to create an \code{image} object with a desired color and size.
}
\examples{
if (interactive()) {
image_create("red")
image_create("#009E73", width = 300, heigth = 100)
}
}
