% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_import_inter}
\alias{image_import_inter}
\title{Import Images Interactively (Windows Only)}
\usage{
image_import_inter()
}
\value{
The function returns the imported images.
}
\description{
This function allows users to interactively select and import images from
their filesystem. It calls \code{\link[=choose.files]{choose.files()}} and \code{\link[=image_import]{image_import()}} and is
specifically designed for Windows operating systems.
}
\details{
This function prompts the user to select one or multiple images from
their filesystem. It only works on Windows operating systems due to
platform-specific file selection mechanisms.
}
\examples{
  if (interactive() && .Platform$OS.type == "windows"){
  library(pliman)
  imgs <- image_import_inter()
  }


}
\seealso{
\code{\link[=image_import]{image_import()}}
}
\keyword{file}
\keyword{import}
\keyword{interactive}
\keyword{windows}
