% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_pallete}
\alias{image_pallete}
\title{Create an image palette}
\usage{
image_pallete(image, npal, nstart = 25)
}
\arguments{
\item{image}{An image object.}

\item{npal}{The number of color palletes.}

\item{nstart}{How many random sets from \code{npal} should be chosen?}
}
\description{
Creates image palettes by applying the k-means algorithm to the RGB values.
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
pal <- image_pallete(img, 2)
image_show(pal[[1]])
image_show(pal[[2]])
}
