% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_morphological.R
\name{image_remove_reflection}
\alias{image_remove_reflection}
\title{Remove Specular Light Reflections}
\usage{
image_remove_reflection(img, radius = 5, has_white_bg = TRUE)
}
\arguments{
\item{img}{Grayscale \code{Image} object or numeric matrix.}

\item{radius}{Radius of structuring window for specular peak removal (default = 5).}

\item{has_white_bg}{Logical indicating whether the background is white/light (default = TRUE).}
}
\value{
A filtered \code{Image} object or matrix.
}
\description{
Removes specular light reflections from grayscale index images prior to thresholding.
}
