% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_rgb}
\alias{image_rgb}
\title{Produces an RGB plot of an image}
\usage{
image_rgb(image, facet = TRUE)
}
\arguments{
\item{image}{An image object.}

\item{facet}{Shows RGB values as a facet plot? Defaults to \code{TRUE}.}
}
\description{
Produces an RGB plot of an image
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
dim(img)
img2 <- image_resize(img, 50)
image_rgb(img2)
}
