% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_square}
\alias{image_square}
\title{Squares an image}
\usage{
image_square(img, plot = TRUE, ...)
}
\arguments{
\item{img}{An \code{image} object.}

\item{plot}{Plots the extended image? defaults to \code{FALSE}.}

\item{...}{Further arguments passed on to \code{\link[=image_expand]{image_expand()}}.}
}
\value{
The modified \code{image} object.
}
\description{
Converts a rectangular image into a square image by expanding the
rows/columns using \code{\link[=image_expand]{image_expand()}}.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("soybean_touch.jpg")
dim(img)
square <- image_square(img)
dim(square)
}
}
