% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{mean.image}
\alias{mean.image}
\title{Mean method for \code{image} objects}
\usage{
\method{mean}{image}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \code{image} object.}

\item{...}{Additional arguments.}

\item{na.rm}{Logical; should NAs be removed?}
}
\description{
Mean method for \code{image} objects
}
