% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_disease_byl.R
\name{measure_disease_byl}
\alias{measure_disease_byl}
\title{Performs plant disease measurements by leaf}
\usage{
measure_disease_byl(
  img,
  index = "B",
  index_lb = "B",
  index_dh = "NGRDI",
  lower_size = NULL,
  watershed = TRUE,
  invert = FALSE,
  fill_hull = FALSE,
  opening = c(10, 0),
  closing = c(0, 0),
  filter = c(0, 0),
  erode = c(0, 0),
  dilate = c(0, 0),
  threshold = "Otsu",
  extension = NULL,
  tolerance = NULL,
  object_size = "large",
  img_healthy = NULL,
  img_symptoms = NULL,
  plot = TRUE,
  save_image = FALSE,
  dir_original = NULL,
  dir_processed = NULL,
  pattern = NULL,
  parallel = FALSE,
  workers = NULL,
  show_features = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{img}{The image to be analyzed.}

\item{index}{A character value specifying the target mode for
conversion to binary to segment the leaves from background. Defaults to "B"
(blue). See \code{\link[=image_index]{image_index()}} for more details. Personalized indexes can be
informed as, e.g., \verb{index = "R*G/B}.}

\item{index_lb}{The index used to segment the foreground (e.g., leaf) from
the background. If not declared, the entire image area (pixels) will be
considered in the computation of the severity.}

\item{index_dh}{The index used to segment diseased from healthy tissues when
\code{img_healthy} and \code{img_symptoms} are not declared. Defaults to \code{"GLI"}. See
\code{\link[=image_index]{image_index()}} for more details.}

\item{lower_size}{To prevent dust from affecting object segmentation, objects
with lesser than \verb{10\%} of the mean of all objects are removed. . One can
set a known area or use \code{lower_limit = 0} to select all objects (not
advised).}

\item{watershed}{If \code{TRUE} (default) performs watershed-based object detection.}

\item{invert}{Inverts the binary image if desired.}

\item{fill_hull}{Fill holes in the binary image? Defaults to \code{FALSE}.}

\item{opening, closing, filter, erode, dilate}{\strong{Morphological operations (brush
size)}, applied in hierarchical order.}

\item{threshold}{The threshold method (\code{"Otsu"}, \code{"Adaptive"}, or numeric).}

\item{extension}{Neighborhood radius for watershed segmentation.}

\item{tolerance}{Minimum height for watershed segmentation.}

\item{object_size}{The size of the object (\code{"small"}, \code{"medium"}, \code{"large"}, \code{"elarge"}).
Used to automatically set up \code{tolerance} and \code{extension} parameters.}

\item{img_healthy}{A color palette of healthy tissues.}

\item{img_symptoms}{A color palette of lesioned tissues.}

\item{plot}{Show image after processing?}

\item{save_image}{Save the image after processing? The image is saved in the
current working directory named as \verb{proc_*} where \code{*} is the image name
given in \code{img}.}

\item{dir_original, dir_processed}{The directory containing the original and
processed images. Defaults to \code{NULL}. In this case, the function will
search for the image \code{img} in the current working directory. After
processing, when \code{save_image = TRUE}, the processed image will be also
saved in such a directory. It can be either a full path, e.g.,
\code{"C:/Desktop/imgs"}, or a subfolder within the current working directory,
e.g., \code{"/imgs"}.}

\item{pattern}{A pattern of file name used to identify images to be
processed. For example, if \code{pattern = "im"} all images that the name
matches the pattern (e.g., img1.-, image1.-, im2.-) will be analyzed.
Providing any number as pattern (e.g., \code{pattern = "1"}) will select
images that are named as 1.-, 2.-, and so on.}

\item{parallel}{Processes the images asynchronously (in parallel) in separate
R sessions running in the background on the same machine. It may speed up
the processing time, especially when \code{pattern} is used is informed. The
number of sections is set up to 30\% of available cores.}

\item{workers}{A positive numeric scalar or a function specifying the maximum
number of parallel processes that can be active at the same time.}

\item{show_features}{If \code{TRUE} returnS the lesion features such as number,
area, perimeter, and radius. Defaults to \code{FALSE}.}

\item{verbose}{If \code{TRUE} (default) shows progress summary in console.}

\item{...}{Additional arguments passed on to \code{\link[=measure_disease]{measure_disease()}}.}
}
\value{
\itemize{
\item A list with the following objects:
\itemize{
\item \code{severity} A data frame with the percentage of healthy and symptomatic
areas for each leaf in the image(s).
\item \code{shape},\code{statistics} If \code{show_features = TRUE} is used, returns the shape
(area, perimeter, etc.) for each lesion and a summary statistic of the
results.
}
}
}
\description{
Computes the percentage of symptomatic leaf area using color palettes or RGB
indexes \code{by} each \code{l}eaf of an image. This allows, for example, processing
replicates of the same treatment  and obtaining the results for each
replication with a single image. To do that, leaf samples are first splitten
with \code{\link[=object_split]{object_split()}} and then, \code{\link[=measure_disease]{measure_disease()}} is applied to the list of
leaves.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("mult_leaves.jpg", plot = TRUE)
sev <-
 measure_disease_byl(img = img,
                     index_lb = "B",
                     index_dh = "NGRDI",
                     workers = 2)
sev$severity
}


}
