% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_injury.R
\name{measure_injury}
\alias{measure_injury}
\title{Measures Injury in Images}
\usage{
measure_injury(
  img = NULL,
  pattern = NULL,
  index = "GRAY",
  threshold = "Otsu",
  invert = FALSE,
  opening = 5,
  closing = FALSE,
  filter = FALSE,
  erode = FALSE,
  dilate = FALSE,
  filter_order = c("erode", "dilate", "opening", "closing", "filter", "fill_hull"),
  plot = TRUE,
  dir_original = NULL,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{img}{The image to be analyzed.}

\item{pattern}{A pattern of file name used to identify images to be imported.}

\item{index}{A character value specifying the index for conversion to binary image.}

\item{threshold}{The threshold method (\code{"Otsu"}, \code{"Adaptive"}, or numeric).}

\item{invert}{Inverts the binary image if desired.}

\item{opening, closing, filter, erode, dilate}{\strong{Morphological operations (brush
size)}, applied in hierarchical order.}

\item{filter_order}{Hierarchical order of morphological operations.}

\item{plot}{Show image after processing?}

\item{dir_original}{The directory containing the original and processed
images. Defaults to NULL. In this case, the function will search for the
image img in the current working directory.}

\item{parallel}{If \code{TRUE} processes the images asynchronously in parallel.}

\item{workers}{Number of parallel processes.}

\item{verbose}{If \code{TRUE} (default) shows progress summary in console.}
}
\value{
A numeric value representing the injury percentage for a single
image, or a data frame with injury percentages for batch processing.
}
\description{
The \code{measures_injury} function calculates the percentage of injury in images
by performing binary segmentation and identifying lesions. It processes
either a single image or a batch of images specified by a pattern in a
directory.
}
\details{
The function processes each image by reading it, applying binary
segmentation to detect lesions, filling the segmented areas, calculating
the injury percentage, and optionally saving the processed image with
highlighted lesions. In batch mode, it uses the provided pattern to
identify images in the specified directory and can utilize parallel
processing for efficiency.
}
