% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mosaic.R
\name{mosaic_chm_mask}
\alias{mosaic_chm_mask}
\title{Apply a height mask to CHM data}
\usage{
mosaic_chm_mask(chm, lower, upper = NULL)
}
\arguments{
\item{chm}{A list or object containing CHM data with a \code{height} field,
typically in a \code{chm} slot.}

\item{lower}{A numeric value specifying the lower height threshold. All
heights greater than this value are retained.}

\item{upper}{An optional numeric value specifying the upper height threshold.
If provided, only heights between \code{lower} and \code{upper} are retained.}
}
\value{
An \code{SpatRaster} (mask).
}
\description{
This function applies a mask to a Canopy Height Model (CHM) by selecting
areas with heights that are above a specified \code{lower} threshold and,
optionally, below an \code{upper} threshold.
}
