% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mosaic.R
\name{mosaic_plot_rgb}
\alias{mosaic_plot_rgb}
\title{A wrapper around terra::plotRGB()}
\usage{
mosaic_plot_rgb(mosaic, ...)
}
\arguments{
\item{mosaic}{SpatRaster}

\item{...}{Further arguments passed on to \code{\link[terra:plotRGB]{terra::plotRGB()}}.}
}
\value{
A \code{NULL} object
}
\description{
Plot the RGB of a SpatRaster
}
