% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_objects.R
\name{object_split}
\alias{object_split}
\title{Splits objects from an image into multiple images}
\usage{
object_split(
  img,
  index = "NB",
  lower_size = NULL,
  watershed = TRUE,
  invert = FALSE,
  fill_hull = FALSE,
  opening = 3,
  closing = FALSE,
  filter = FALSE,
  erode = FALSE,
  dilate = FALSE,
  threshold = "Otsu",
  extension = NULL,
  tolerance = NULL,
  object_size = "medium",
  edge = 3,
  remove_bg = FALSE,
  plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{img}{The image to be analyzed.}

\item{index}{A character value specifying the index for conversion to binary image.}

\item{lower_size}{Plant images often contain dirt and dust. To prevent dust from
affecting the image analysis, objects with lesser than 10\% of the mean of all objects
are removed. Set \code{lower_limit = 0} to keep all the objects.}

\item{watershed}{If \code{TRUE} (default) performs watershed-based object detection.}

\item{invert}{Inverts the binary image if desired.}

\item{fill_hull}{Fill holes in the binary image? Defaults to \code{FALSE}.}

\item{opening, closing, filter, erode, dilate}{\strong{Morphological operations (brush
size)}, applied in hierarchical order.}

\item{threshold}{The threshold method (\code{"Otsu"}, \code{"Adaptive"}, or numeric).}

\item{extension}{Neighborhood radius for watershed segmentation.}

\item{tolerance}{Minimum height for watershed segmentation.}

\item{object_size}{The size of the object (\code{"small"}, \code{"medium"}, \code{"large"}, \code{"elarge"}).
Used to automatically set up \code{tolerance} and \code{extension} parameters.}

\item{edge}{The number of pixels to be added in the edge of the segmented
object. Defaults to 5.}

\item{remove_bg}{If \code{TRUE}, the pixels that are not part of objects are
converted to white.}

\item{plot}{Show image after processing?}

\item{verbose}{If \code{TRUE} (default) shows progress summary in console.}

\item{...}{Additional arguments passed on to \code{\link[=image_combine]{image_combine()}}}
}
\value{
A list of objects of class \code{image}.
}
\description{
Using threshold-based segmentation, objects are first isolated from
background. Then, a new image is created for each single object. A list of
images is returned.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("la_leaves.jpg", plot = TRUE)
imgs <- object_split(img) # set to NULL to use 50\% of the cores
}

}
\seealso{
\code{\link[=analyze_objects]{analyze_objects()}}, \code{\link[=image_binary]{image_binary()}}
}
