% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{Ops.image}
\alias{Ops.image}
\title{Group generic Ops method for \code{image} objects}
\usage{
\method{Ops}{image}(e1, e2)
}
\arguments{
\item{e1}{Left operand.}

\item{e2}{Right operand.}
}
\description{
Group generic Ops method for \code{image} objects
}
