% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_neural.R
\name{pliman_onnx_dir}
\alias{pliman_onnx_dir}
\title{Directory for ONNX Runtime C++ Shared Libraries}
\usage{
pliman_onnx_dir(dir = NULL)
}
\arguments{
\item{dir}{Optional custom directory path.}
}
\value{
A character string with the path to the bin directory.
}
\description{
Gets or sets the local directory where the official Microsoft ONNX Runtime
C++ shared libraries (\code{onnxruntime.dll}, \code{libonnxruntime.so}, or \code{libonnxruntime.dylib})
are saved.
}
\examples{
pliman_onnx_dir()
}
