% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_neural.R
\name{pliman_onnx_lib_path}
\alias{pliman_onnx_lib_path}
\title{Locate ONNX Runtime C++ Shared Library}
\usage{
pliman_onnx_lib_path(engine = c("cpu", "gpu"))
}
\arguments{
\item{engine}{Execution engine: \code{"cpu"} (default) or \code{"gpu"}.}
}
\value{
Character path to the dynamic library or \code{NULL} if not found.
}
\description{
Finds the local path to the ONNX Runtime dynamic library (\code{onnxruntime.dll} on Windows,
\code{libonnxruntime.so} on Linux, \code{libonnxruntime.dylib} on macOS), or \code{onnxruntime-directml.dll}
when \code{engine = "gpu"} on Windows.
}
