% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pliman_viewer}
\alias{pliman_viewer}
\title{Global option for controlling the viewer in pliman package}
\description{
Users can set the value of this option using \code{options("pliman_viewer", value)}.
The default value is "base". Use "mapview" to allow image to be
plotted/edited using the R packages mapview and mapedit
}
