% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{plot.image_index}
\alias{plot.image_index}
\title{Plots an \code{image_index} object
Produces an histogram of an \code{image_index} object
@name image_index}
\usage{
\method{plot}{image_index}(x, facet = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{image_index}.}

\item{facet}{Shows RGB values as a facet plot? Defaults to \code{TRUE}.}

\item{...}{Currently not used}
}
\value{
A \code{ggplot} object containing the distribution of the pixels for each
index.
}
\description{
Plots an \code{image_index} object
Produces an histogram of an \code{image_index} object
@name image_index
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
img2 <- image_resize(img, 50)
ind <- image_index(img2)
plot(ind)
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
