% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_rgb.R
\name{plot.objects_rgb}
\alias{plot.objects_rgb}
\title{Plots an \code{objects_rgb} object
Produces an histogram of an \code{objects_rgb} object
@name objects_rgb}
\usage{
\method{plot}{objects_rgb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{objects_rgb}.}

\item{...}{Currently not used}
}
\value{
A \code{ggplot} object containing the distribution of the pixels for each
index.
}
\description{
Plots an \code{objects_rgb} object
Produces an histogram of an \code{objects_rgb} object
@name objects_rgb
}
\examples{
\donttest{
library(pliman)

img <- image_import(image_pliman("soy_green.jpg"))
# Segment the foreground (grains) using the normalized blue index
# Shows the average value of the blue index in each object
rgb <- objects_rgb(img)
plot(rgb)
}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
