% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{plot_line_segment}
\alias{plot_line_segment}
\title{Plot Detected Line Segments}
\usage{
plot_line_segment(x, col = "red", lwd = 1)
}
\arguments{
\item{x}{A list returned by \code{\link[=image_line_segment]{image_line_segment()}}, containing detected line segments.}

\item{col}{The color of lines}

\item{lwd}{The width of lines. Defaults to 1}
}
\value{
No return value. The function adds line segments to an existing plot.
}
\description{
Plots the detected line segments from the output of \code{\link[=image_line_segment]{image_line_segment()}}.
Each segment is drawn as a red line on the existing plot.
}
\examples{
library(pliman)

}
