% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_polygon.R
\name{poly_base_apex_angle}
\alias{poly_base_apex_angle}
\title{Calculate the apex and base angles of an object}
\usage{
poly_base_apex_angle(x, percentiles = c(0.25, 0.75))
}
\arguments{
\item{x}{A logical matrix or matrix of coordinates representing the object}

\item{percentiles}{A numeric vector of two percentiles between 0 and 1
indicating the height of the points from the top to the bottom. The function
calculates the apex angle between the two percentiles and the base angle
between the lowest point and the highest point.}
}
\value{
A list containing the apex angle, base angle, apex coordinates, and
base coordinates.
}
\description{
This function calculates the apex and base angles of an object. It takes as
input either a logical matrix (binarized image) or a matrix of coordinates.
The function returns the apex angle, base angle, and the coordinates of the
apex and base as a list.
}
\examples{
library(pliman)
img <- image_pliman("sev_leaf.jpg")
bin <- image_binary(img, "NB", plot = FALSE)[[1]]
angls <- poly_base_apex_angle(bin)
plot(img)
points(angls$apex_coords, pch = 16, cex = 2, col = "red")
points(angls$base_coords, pch = 16, cex = 2, col = "red")
angls
}
