% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{print.image}
\alias{print.image}
\title{Print method for \code{image} objects}
\usage{
\method{print}{image}(x, ...)
}
\arguments{
\item{x}{An \code{image} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{image} objects
}
