% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_video.R
\name{print.pliman_video_frames}
\alias{print.pliman_video_frames}
\title{Print method for pliman_video_frames}
\usage{
\method{print}{pliman_video_frames}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pliman_video_frames}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for pliman_video_frames
}
