% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_root.R
\name{print.root_analysis}
\alias{print.root_analysis}
\title{Print method for root_analysis objects}
\usage{
\method{print}{root_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{root_analysis} object.}

\item{...}{Ignored.}
}
\description{
Print method for root_analysis objects
}
