% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{random_color}
\alias{random_color}
\title{Random built-in color names}
\usage{
random_color(n = 1, distinct = FALSE)
}
\arguments{
\item{n}{The number of color names. Defaults to 1.}

\item{distinct}{Logical indicating if the colors returned should all be
distinct. Defaults to \code{FALSE}.}
}
\value{
A character vector of color names
}
\description{
Randomly chooses single or multiple built-in color names which R knows about.
See more at \code{\link[grDevices:colors]{grDevices::colors()}}
}
\examples{
library(pliman)
random_color(n = 3)
}
