% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_measures.R
\name{report_index}
\alias{report_index}
\title{Reports for an object index}
\usage{
report_index(object, index, cut_point, select_higher = FALSE)
}
\arguments{
\item{object}{An object computed with \code{\link[=analyze_objects]{analyze_objects()}}.}

\item{index}{The index desired, e.g., "B". Note that these value must match the index(es) used in}

\item{cut_point}{The cut point}

\item{select_higher}{If \code{FALSE} (default) selects the objects with \code{index}
smaller than the \code{cut_point}. Use \code{select_higher = TRUE} to select the
objects with \code{index} higher than \code{cut_point}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{ids} The identification of selected objects.
\item \code{report} A data frame with the following columns
\itemize{
\item \code{n} The number of objects.
\item \code{nsel} The number of selected objects.
\item \code{prop} The proportion of objects selected.
\item \code{mean_index_sel}, and \code{mean_index_nsel} The mean value of \code{index} for the
selected and non-selected objects, respectively.
}
}
}
\description{
Performs a report of the object index when \code{object_index} argument is used in
\code{analyze_objects()}. By using a cut point, the proportion and mean values for
the selected objects are returned.
}
\examples{
soy <- image_pliman("soy_green.jpg")
anal <- analyze_objects(soy, object_index = "G")
plot_measures(anal, measure = "G")

report_index(anal, index = "G", cut_point = 0.5)
}
