% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{rgb_to_hsv}
\alias{rgb_to_hsv}
\title{Color spaces}
\usage{
rgb_to_hsv(image)
}
\arguments{
\item{image}{An image object.}
}
\value{
A list containing the image in the new color space.
}
\description{
Convert RGB to LAB color space.
}
\examples{
library(pliman)
img <- image_pliman("sev_leaf.jpg")
img2 <- rgb_to_hsv(img)
image_combine(img, img2)
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
