% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{run_app}
\alias{run_app}
\title{Shiny UI for pliman package}
\usage{
run_app()
}
\description{
run_app calls plimanshiny::run_app() that will starts the Shiny interface of
the pliman package
}
\examples{
if(interactive()){
library(pliman)
run_app()
}


}
