% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_pliman_viewer}
\alias{set_pliman_viewer}
\title{Set the value of the pliman_viewer option}
\usage{
set_pliman_viewer(value)
}
\arguments{
\item{value}{The value to be set for the pliman_viewer option.}
}
\description{
Sets the value of the pliman_viewer option used in the package.
}
