% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_wd_here}
\alias{set_wd_here}
\title{Set the Working Directory quicky}
\usage{
set_wd_here(path = NULL)
}
\arguments{
\item{path}{Path components below the project root. Defaults to \code{NULL}. This means that
the directory will be set to the path of the file.}
}
\value{
A message showing the current working directory
}
\description{
It sets the working directory to the path of the current script.
}
\examples{

\dontrun{
set_wd_here()
}
}
