% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{[.image}
\alias{[.image}
\title{Subsetting method for \code{image} objects}
\usage{
\method{[}{image}(x, i, j, k, ..., drop = FALSE)
}
\arguments{
\item{x}{An \code{image} object.}

\item{i}{Row/X index or missing.}

\item{j}{Column/Y index or missing.}

\item{k}{Channel index or missing.}

\item{...}{Additional arguments.}

\item{drop}{Logical; whether to drop dimensions.}
}
\description{
Subsetting method for \code{image} objects
}
