% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{Summary.image}
\alias{Summary.image}
\title{Group generic Summary method for \code{image} objects (min, max, range, sum, prod, any, all)}
\usage{
\method{Summary}{image}(..., na.rm = FALSE)
}
\arguments{
\item{...}{\code{image} objects or numbers.}

\item{na.rm}{Logical; should NAs be removed?}
}
\description{
Group generic Summary method for \code{image} objects (min, max, range, sum, prod, any, all)
}
