% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_class.R
\name{to_numeric}
\alias{to_numeric}
\alias{to_raw}
\title{Convert \code{image} between raw (0..255 uint8) and numeric (0..1 double)}
\usage{
to_numeric(x)

to_raw(x)
}
\arguments{
\item{x}{An \code{image} object.}
}
\value{
An \code{image} object converted to numeric storage.
}
\description{
Convert \code{image} between raw (0..255 uint8) and numeric (0..1 double)
}
