% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indexes.R
\name{utils_indexes}
\alias{utils_indexes}
\alias{pliman_indexes}
\alias{pliman_indexes_eq}
\alias{pliman_indexes_rgb}
\alias{pliman_indexes_me}
\title{Utilities for image indexes}
\usage{
pliman_indexes()

pliman_indexes_eq()

pliman_indexes_rgb()

pliman_indexes_me()
}
\description{
\itemize{
\item \code{pliman_indexes()}: Get all the available indexes in pliman.
\item \code{pliman_indexes_rgb()}: Get all the RGB-based available indexes in pliman.
\item \code{pliman_indexes_me()}: Get all the multispectral available indexes in pliman.
\item \code{pliman_indexes_eq()}: Get the equations of the available indexes.
}
}
