% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_measures.R
\name{utils_stats}
\alias{utils_stats}
\alias{mean_list}
\alias{sd_list}
\alias{max_list}
\alias{min_list}
\title{These functions applies common statistics to a list of objects, returning a
numeric vector.}
\usage{
mean_list(x, ...)

sd_list(x, ...)

max_list(x, ...)

min_list(x, ...)
}
\arguments{
\item{x}{A data.frame or matrix with numeric values.}

\item{...}{Further arguments passed on to the R base function (e.g, mean(),
sd(), etc.)}
}
\value{
A numeric vector.
}
\description{
These functions applies common statistics to a list of objects, returning a
numeric vector.
}
\examples{
mean_list(list(a = 1:10, b = 2:20))
}
