% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_methods.R
\name{view}
\alias{view}
\title{Generic function for interactive viewing of images and spatial objects}
\usage{
view(x, ...)
}
\arguments{
\item{x}{An image or spatial object to view.}

\item{...}{Additional arguments passed to specific view methods.}
}
\description{
Generic function for interactive viewing of images and spatial objects
}
