% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shp.R
\name{analyze_objects_shp}
\alias{analyze_objects_shp}
\title{Analyzes objects using shapefiles}
\usage{
analyze_objects_shp(
  img,
  shapefile = NULL,
  rows = NULL,
  cols = NULL,
  show_image = TRUE,
  object_size = "large",
  ...
)
}
\arguments{
\item{img}{An \code{Image} object}

\item{shapefile}{(Optional) An object created with \code{\link[=image_shp]{image_shp()}}. If \code{NULL}
(default), both \code{rows} and \code{cols} must be declared.}

\item{rows, cols}{The number of rows and columns to generate the shapefile
when \code{shapefile} is not declared.}

\item{show_image}{Shows the processed images? Defaults to \code{TRUE}.}

\item{object_size}{Argument to control control the watershed segmentation.
See \code{\link[=analyze_objects]{analyze_objects()}} for more details.}

\item{...}{Aditional arguments passed on to \link{analyze_objects}.}
}
\value{
An object of class \code{anal_obj}. See more details in the \code{Value}
section of \code{\link[=analyze_objects]{analyze_objects()}}.
}
\description{
This function calls \code{\link[=analyze_objects]{analyze_objects()}} in each image polygon of a shapefile
object generated with generated with \code{\link[=image_shp]{image_shp()}} and bind the results into
read-ready data frames.
}
\examples{
if(interactive()){
library(pliman)
flax <- image_pliman("flax_leaves.jpg", plot = TRUE)
res <- analyze_objects_shp(flax, rows = 3, cols = 5)
}
}
