% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erode_external_cpp}
\alias{erode_external_cpp}
\title{Erosão externa rápida (preserva borda interna de objetos fechados)}
\usage{
erode_external_cpp(img, raio = 3L, forma = "disc")
}
\description{
Erosão externa rápida (preserva borda interna de objetos fechados)
}
