% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erosao_conservativa}
\alias{erosao_conservativa}
\title{Fast conservative erosion pipeline}
\usage{
erosao_conservativa(img, raio_erosao = 3L, raio_dilatacao = 3L, forma = "disc")
}
\arguments{
\item{img}{A logical matrix representing the binary image.}

\item{raio_erosao}{Radius for erosion step (integer >= 1).}

\item{raio_dilatacao}{Radius for dilation step (integer >= 1).}

\item{forma}{Shape of the structuring element ("disc" or "square").}
}
\description{
Fast conservative erosion pipeline
}
