% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_uuid}
\alias{get_uuid}
\title{Extract UUID from filenames}
\usage{
get_uuid(filename)
}
\arguments{
\item{filename}{A character vector containing filenames or strings}
}
\value{
A character vector with extracted UUIDs (or NA if none found)
}
\description{
This function extracts a UUID (Universal Unique Identifier) from the
filename, using a regular expression that specifically identifies the
standard UUID format.
}
\examples{
library(pliman)
file <- "Grãos - contagem_f68bca60-c8cf-4272-9448-3f28891a97cd.jpg"
file2 <- "Grãos - contagem_f68bca60-c8cf-4272-9448-3f8891a97cd.jpg"
get_uuid(file)
}
