% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shp.R
\name{image_align}
\alias{image_align}
\title{Aligns an \code{Image} object by hand}
\usage{
image_align(img, align = c("vertical", "horizontal"), plot = TRUE)
}
\arguments{
\item{img}{An \code{Image} object}

\item{align}{The desired alignment. Either \code{"vertical"} (default) or
\code{"horizontal"}.}

\item{plot}{Plots the aligned image? Defaults to \code{TRUE}.}
}
\value{
The \code{img} aligned
}
\description{
\code{\link[=image_rotate]{image_rotate()}} rotate an image given an informed angle. Here, users will
need to draw a line along the y axis that corresponds to the alignment of the
objects (e.g., field plots). By default, the aligment will be to the
vertical, which means that if the drawed line have an angle < 90º parallel to
the x axis, the rotation angle wil be negative (anticlocwise rotation). If
the drawed line have an angle > 90º along the x axis, the rotation angle wil
be positive (clocwise rotation). If the aligment is horizontal, the image
will be rotated to align the drawed line paralell to the x axis assuming the
shortest angle (left or right rotation if the angle is < 90º or > 90º,
respectively)
}
\examples{
if(interactive()){
library(pliman)
flax <- image_pliman("flax_leaves.jpg", plot = TRUE)
aligned <- image_align(flax)
}
}
