% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_binary}
\alias{image_binary}
\title{Creates a binary image}
\usage{
image_binary(
  img,
  index = "R",
  r = 1,
  g = 2,
  b = 3,
  re = 4,
  nir = 5,
  return_class = "image",
  threshold = c("Otsu", "adaptive"),
  k = 0.15,
  windowsize = NULL,
  has_white_bg = FALSE,
  resize = FALSE,
  fill_hull = FALSE,
  max_size = NULL,
  remove_reflection = FALSE,
  erode = FALSE,
  dilate = FALSE,
  opening = FALSE,
  closing = FALSE,
  filter = FALSE,
  filter_order = c("erode", "dilate", "opening", "closing", "filter", "fill_hull"),
  invert = FALSE,
  return_exact = FALSE,
  plot = TRUE,
  nrow = NULL,
  ncol = NULL,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{img}{An image object.}

\item{index}{A character value (or a vector of characters) specifying the
target mode for conversion to binary image. See the available indexes with
\code{\link[=pliman_indexes]{pliman_indexes()}} and \code{\link[=image_index]{image_index()}} for more details.}

\item{r, g, b, re, nir}{The red, green, blue, red-edge, and near-infrared bands
of the image, respectively. Defaults to 1, 2, 3, 4, and 5, respectively. If
a multispectral image is provided (5 bands), check the order of bands,
which are frequently presented in the 'BGR' format.}

\item{return_class}{The class of object to be returned. If \verb{"terra} returns a
SpatRaster object with the number of layers equal to the number of indexes
computed. If \code{"ebimage"} (default) returns a list of \code{image} objects, where
each element is one index computed.}

\item{threshold}{The theshold method to be used.
\itemize{
\item By default (\code{threshold = "Otsu"}), a threshold value based
on Otsu's method is used to reduce the grayscale image to a binary image. If
a numeric value is informed, this value will be used as a threshold.
\item If \code{threshold = "adaptive"}, adaptive thresholding (Shafait et al. 2008)
is used, and will depend on the \code{k} and \code{windowsize} arguments.
\item If any non-numeric value different than \code{"Otsu"} and \code{"adaptive"} is used,
an iterative section will allow you to choose the threshold based on a
raster plot showing pixel intensity of the index.
}}

\item{k}{a numeric in the range 0-1. when \code{k} is high, local threshold
values tend to be lower. when \code{k} is low, local threshold value tend to be
higher.}

\item{windowsize}{windowsize controls the number of local neighborhood in
adaptive thresholding. By default it is set to \code{1/3 * minxy}, where
\code{minxy} is the minimum dimension of the image (in pixels).}

\item{has_white_bg}{Logical indicating whether a white background is present.
If \code{TRUE}, pixels that have R, G, and B values equals to 1 will be
considered as \code{NA}. This may be useful to compute an image index for
objects that have, for example, a white background. In such cases, the
background will not be considered for the threshold computation.}

\item{resize}{Resize the image before processing? Defaults to \code{FALSE}. Use a
numeric value as the percentage of desired resizing. For example, if
\code{resize = 30}, the resized image will have 30\% of the size of original
image.}

\item{fill_hull}{Fill holes in the objects? Defaults to \code{FALSE}.}

\item{erode, dilate, opening, closing, filter}{\strong{Morphological operations (brush size)}
\itemize{
\item \code{dilate} puts the mask over every background pixel, and sets it to
foreground if any of the pixels covered by the mask is from the foreground.
\item \code{erode} puts the mask over every foreground pixel, and sets it to
background if any of the pixels covered by the mask is from the background.
\item \code{opening} performs an erosion followed by a dilation. This helps to
remove small objects while preserving the shape and size of larger objects.
\item \code{closing} performs a dilatation followed by an erosion. This helps to
fill small holes while preserving the shape and size of larger objects.
\item \code{filter} performs median filtering in the binary image. Provide a positive
integer > 1 to indicate the size of the median filtering. Higher values are
more efficient to remove noise in the background but can dramatically impact
the perimeter of objects, mainly for irregular perimeters such as leaves
with serrated edges.
}

Hierarchically, the operations are performed as opening > closing > filter.
The value declared in each argument will define the brush size.}

\item{filter_order}{A character vector indicating the order in which morphological
operations are applied. Defaults to \code{c("erode", "dilate", "opening", "closing", "filter", "fill_hull")}.}

\item{invert}{Inverts the binary image, if desired.}

\item{return_exact}{Logical. If \code{TRUE}, exact morphological reconstruction via
watershed-partitioned seed matching is performed when \code{opening > 0}.}

\item{plot}{Show image after processing?}

\item{nrow, ncol}{The number of rows or columns in the plot grid. Defaults to
\code{NULL}, i.e., a square grid is produced.}

\item{parallel}{Processes the images asynchronously (in parallel) in separate
R sessions running in the background on the same machine. It may speed up
the processing time when \code{image} is a list. The number of sections is set
up to 70\% of available cores.}

\item{workers}{A positive numeric scalar or a function specifying the maximum
number of parallel processes that can be active at the same time.}

\item{verbose}{If \code{TRUE} (default) a summary is shown in the console.}
}
\value{
A list containing binary images. The length will depend on the number
of indexes used.
}
\description{
Reduce a color, color near-infrared, or grayscale images to a binary image
using a given color channel (red, green blue) or even color indexes. The
Otsu's thresholding method (Otsu, 1979) is used to automatically perform
clustering-based image thresholding.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("soybean_touch.jpg")
image_binary(img, index = c("R, G"))
}

}
\references{
Otsu, N. 1979. Threshold selection method from gray-level histograms. IEEE
Trans Syst Man Cybern SMC-9(1): 62–66. \doi{10.1109/tsmc.1979.4310076}

Shafait, F., D. Keysers, and T.M. Breuel. 2008. Efficient implementation of
local adaptive thresholding techniques using integral images. Document
Recognition and Retrieval XV. SPIE. p. 317–322 \doi{10.1117/12.767755}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
