% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_binary}
\alias{image_binary}
\title{Creates a binary image}
\usage{
image_binary(image, channel = "blue", invert = FALSE)
}
\arguments{
\item{image}{An image object.}

\item{channel}{A character value specifying the target mode for conversion to
binary image. One of \code{gray}, \code{grey}, \code{red}, \code{green}, or \code{blue}. Defaults to
\code{"blue"}.}

\item{invert}{Inverts the binary image, if desired.}
}
\description{
Reduce a color or grayscale image to a binary image using a given color
channel. Otsu's thresholding method (Otsu, 1979) is used to automatically
perform clustering-based image thresholding.
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
image_show(image_binary(img, "red"))
image_show(image_binary(img, "red"))
}
\references{
Nobuyuki Otsu, "A threshold selection method from gray-level
histograms". IEEE Trans. Sys., Man., Cyber. 9 (1): 62-66. 1979.
\doi{10.1109/TSMC.1979.4310076}
}
