% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_combine}
\alias{image_combine}
\title{Combines images to a grid}
\usage{
image_combine(..., nrow = NULL, ncol = NULL)
}
\arguments{
\item{...}{a comma-separated list of image objects.}

\item{nrow, ncol}{The number of rows or columns in the plot grid. Defaults to
\code{NULL}, i.e., a square grid is produced.}
}
\description{
Combines several images to a grid
}
\examples{
library(pliman)
img1 <- image_import(system.file("tmp_images", "sev1.png", package = "pliman"))
img2 <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
image_combine(img1, img2)
}
