% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_combine}
\alias{image_combine}
\title{Combines images to a grid}
\usage{
image_combine(..., nrow = NULL, ncol = NULL)
}
\arguments{
\item{...}{a comma-separated name of image objects or a list containing image
objects.}

\item{nrow, ncol}{The number of rows or columns in the plot grid. Defaults to
\code{NULL}, i.e., a square grid is produced.}
}
\value{
A grid with the images in \code{...}
}
\description{
Combines several images to a grid
}
\examples{
library(pliman)
img1 <- image_import(image_pliman("sev_leaf.jpg"))
img2 <- image_import(image_pliman("sev_leaf_nb.jpg"))
image_combine(img1, img2)
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
