% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_correction_pick}
\alias{image_correction_pick}
\title{Interactive Color Correction}
\usage{
image_correction_pick(img, known_colors, color_chips = 24)
}
\arguments{
\item{img}{An \code{Image} object to be corrected.}

\item{known_colors}{A \code{data.frame} containing the target reference values.
Must contain the columns \code{R}, \code{G}, and \code{B} (normalized between 0-1) and
the number of rows must equal \code{color_chips}.}

\item{color_chips}{The number of color patches to be interactively
sampled from the image. Defaults to \code{24}.}
}
\value{
An \code{Image} object with corrected colors.
}
\description{
Interactively calibrates the color of an image by prompting the user to
sample color patches. It first calls \code{pliman::pick_rgb_area()} to pause and
prompt the user to click on \code{color_chips} patches in the image. These
interactively sampled colors are then used as the \emph{observed} colors to
compute the 3rd-order polynomial transformation matrix (K), which maps the
observed colors to the provided \code{known_colors}. This matrix K is then applied
to every pixel in the image.
}
\examples{
if(interactive()){
# colorcheck image available at https://github.com/HarryCWright/PlantSizeClr
library(pliman)

# known values in the following sequence (row, column)
# blue (6, 2)
# red (4, 2)
# yellow (3, 2)
# white (1, 1)
known_colors <- data.frame(
 id = c(1, 2, 3, 4),
 R = c(0.09803922, 0.72941176, 0.96078431, 0.97647059),
 G = c(0.2156863, 0.1019608, 0.8039216, 0.9490196),
 B = c(0.5294118, 0.2000000, 0.0000000, 0.9333333)
)
img <- image_pliman("colorcheck.jpg")
# draw four samples, following the order (blue, red, yellow, white)
 img_cor <- image_correction_pick(img, known_colors, color_chips = 4)
 image_combine(img, img_cor)
}


}
