% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_resize}
\alias{image_resize}
\title{Resize an image object}
\usage{
image_resize(image, rel_size = 100, width, height)
}
\arguments{
\item{image}{An image object.}

\item{rel_size}{The relative size of the resized image. Defaults to 100.}

\item{width, height}{Width and height of the resized image. These arguments
can be missing. In this case, the image is resized according to the
relative size informed in \code{rel_size.}.}
}
\description{
Scales the \code{image} to the desired dimensions.
}
\details{
Image can be resized either by a percent value of the original image or by
explicitly  declaring the desired dimension in the final image. For example,
setting \code{rel_size = 50} to an image of width \verb{1280 x 720}, the new image will
have a size of \verb{640 x 360}. Declare one of \code{width} or \code{height} to enable
proportional resizing by pixel values.
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
dim(img)
dim(image_resize(img, 50))
}
