% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_segment_manual}
\alias{image_segment_manual}
\title{Image segmentation by hand}
\usage{
image_segment_manual(
  image,
  type = c("select", "remove"),
  resize = TRUE,
  edge = 5,
  plot = TRUE
)
}
\arguments{
\item{image}{An \code{Image} object.}

\item{type}{The type of segmentation. By default (\code{type = "select"}) objects
are selected. Use \code{type = "remove"} to remove the selected area from the
image.}

\item{resize}{By default, the segmented object is resized to fill the
original image size. Use \code{resize = FALSE} to keep the segmented object in
the original scale.}

\item{edge}{Number of pixels to add in the edge of the segmented object when
\code{resize = TRUE}. Defaults to 5.}

\item{plot}{Plot the segmented object? Defaults to \code{TRUE}.}
}
\value{
A list with the segmented image and the mask used for segmentation.
}
\description{
Segments image objects 'by hand'. The user will need to pick the perimeter of
the object to be segmented. So, this only works in an interactive section.
}
\examples{
if (interactive()) {
img <- image_pliman("la_leaves.jpg")
seg <- image_segment_manual(img)
plot(seg$mask)

}
}
