% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shp.R
\name{image_shp}
\alias{image_shp}
\title{Construct a shape file from an image}
\usage{
image_shp(
  img,
  rows = 1,
  cols = 1,
  col_line = "red",
  size_line = 2,
  col_text = "red",
  size_text = 1
)
}
\arguments{
\item{img}{An object of class \code{Image}}

\item{rows}{The number of desired rows in the grid. Defaults to \code{1}.}

\item{cols}{The number of desired columns in the grid. Defaults to \code{1}.}

\item{col_line, col_text}{The color of the line/text in the grid. Defaults to
\code{"red"}.}

\item{size_line, size_text}{The size of the line/text in the grid. Defaults to
\code{2.5}.}
}
\value{
A list with \code{row * col} objects containing the plot coordinates.
}
\description{
Creates a list of object coordinates given the desired number of rows and
columns. It starts by selecting 4 points at the corners of objects of
interest in the plot space. Then, given \code{rows} and \code{cols}, a grid is drawn
and the objects' coordinates are returned.
}
\examples{
if(interactive()){
library(pliman)
flax <- image_pliman("flax_leaves.jpg", plot = TRUE)
shape <- image_shp(flax, rows = 3, cols = 5)
}

}
