% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{image_to_mat}
\alias{image_to_mat}
\title{Convert an image to numerical matrices}
\usage{
image_to_mat(image)
}
\arguments{
\item{image}{An image object.
frame? Padrao \code{TRUE}.}
}
\description{
Given an object image, converts it into three matrices (RGB) and a data frame
where each column corresponds to the RGB values.
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
dim(img)
mat <- image_to_mat(img)
dim(mat[[1]])
}
