% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_segmentation.R
\name{image_watershed}
\alias{image_watershed}
\title{Intelligent Morphological Watershed (Spanning Forest + Union-Find)}
\usage{
image_watershed(img, sensitivity = "medium", tolerance = NULL, ext = 1)
}
\arguments{
\item{img}{A binary \code{Image} object or a matrix/array.}

\item{sensitivity}{Text ("low", "medium", "high", "extreme") or Number (e.g. 2.5).
Automatically defines the best parameters for optimal precision.
(Overrides manual adjustment of tolerance).}

\item{tolerance}{Manual Fine-Tuning. If provided, overrides \code{sensitivity}
and directly controls the merging threshold for connected components.}

\item{ext}{Pixel neighborhood radius (default = 1). Specifies the extension radius
used during the regional maxima calculation and watershed flooding.}
}
\value{
A grayscale \code{Image} object containing the labels.
}
\description{
This function performs a morphological watershed segmentation on a binary image.
It implements a highly optimized Spanning Forest algorithm coupled with a Disjoint-Set
(Union-Find) data structure. This approach eliminates the need for slow iterative scans
($O(N^2)$ complexity), allowing the algorithm to execute in milliseconds even when
resolving tens of thousands of distinct objects. It also features a mechanism to merge
highly connected objects, effectively preventing over-segmentation.
}
\examples{
\dontrun{
  library(pliman)
  bin <- image_pliman("soybean_touch.jpg") |> image_binary(index = "B")
  w_med <- image_watershed(bin$B)
}
}
