% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mosaic.R
\name{mosaic_aggregate}
\alias{mosaic_aggregate}
\title{A wrapper around terra::aggregate()}
\usage{
mosaic_aggregate(mosaic, fact = 2, ...)
}
\arguments{
\item{mosaic}{SpatRaster}

\item{fact}{Aggregation factor expressed as number of cells in each direction
(horizontally and vertically). See \code{\link[terra:aggregate]{terra::aggregate()}} for more details.}

\item{...}{Further arguments passed on to \code{\link[terra:aggregate]{terra::aggregate()}}.}
}
\value{
SpatRaster
}
\description{
Aggregate a SpatRaster to create a new SpatRaster with a lower resolution
(larger cells). See \code{\link[terra:aggregate]{terra::aggregate()}} for more details
}
\examples{
library(pliman)
library(terra)
r <- rast()
values(r) <- 1:ncell(r)
r2 <- mosaic_aggregate(r, fact = 10)
opar <- par(no.readonly = TRUE)
par(mfrow=c(1,2))
mosaic_plot(r)
mosaic_plot(r2)
par(opar)
}
