% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mosaic.R
\name{mosaic_hist}
\alias{mosaic_hist}
\title{A wrapper around terra::hist()}
\usage{
mosaic_hist(mosaic, layer, ...)
}
\arguments{
\item{mosaic}{SpatRaster}

\item{layer}{positive integer or character to indicate layer numbers (or
names). If missing, all layers are used}

\item{...}{Further arguments passed on to \code{\link[terra:hist]{terra::hist()}}.}
}
\value{
A \code{NULL} object
}
\description{
Create a histogram of the values of a \code{SpatRaster}.
}
\examples{
library(pliman)
r <- mosaic_input(system.file("ex/elev.tif", package="terra"))
mosaic_hist(r)
}
