% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mosaic.R
\name{mosaic_plot}
\alias{mosaic_plot}
\title{A wrapper around terra::plot()}
\usage{
mosaic_plot(mosaic, ...)
}
\arguments{
\item{mosaic}{SpatRaster}

\item{...}{Further arguments passed on to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
A \code{NULL} object
}
\description{
Plot the values of a SpatRaster
}
\examples{
library(pliman)
r <- mosaic_input(system.file("ex/elev.tif", package="terra"))
mosaic_plot(r)
}
