% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_objects.R
\name{object_export}
\alias{object_export}
\title{Export multiple objects from an image to multiple images}
\usage{
object_export(
  img,
  pattern = NULL,
  dir_original = NULL,
  dir_processed = NULL,
  format = ".jpg",
  squarize = FALSE,
  augment = FALSE,
  times = 12,
  index = "NB",
  lower_size = NULL,
  watershed = FALSE,
  invert = FALSE,
  fill_hull = FALSE,
  opening = 3,
  closing = FALSE,
  filter = FALSE,
  erode = FALSE,
  dilate = FALSE,
  threshold = "Otsu",
  extension = NULL,
  tolerance = NULL,
  object_size = "medium",
  edge = 20,
  remove_bg = FALSE,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{img}{The image to be analyzed.}

\item{pattern}{A pattern of file name used to identify images to be
processed. For example, if \code{pattern = "im"} all images in the current
working directory that the name matches the pattern (e.g., img1.-,
image1.-, im2.-) will be imported and processed. Providing any number as
pattern (e.g., \code{pattern = "1"}) will select images that are named as 1.-,
2.-, and so on. An error will be returned if the pattern matches any file
that is not supported (e.g., img1.pdf).}

\item{dir_original}{The directory containing the original images. Defaults to
\code{NULL}. It can be either a full path, e.g., \code{"C:/Desktop/imgs"}, or a
subfolder within the current working directory, e.g., \code{"/imgs"}.}

\item{dir_processed}{Optional character string indicating a subfolder within the
current working directory to save the image(s). If the folder doesn't
exist, it will be created.}

\item{format}{The format of image to be exported.}

\item{squarize}{Squarizes the image before the exportation? If \code{TRUE},
\code{\link[=image_square]{image_square()}} will be called internally.}

\item{augment}{A logical indicating if exported objects should be augmented using
\code{\link[=image_augment]{image_augment()}}. Defaults to \code{FALSE}.}

\item{times}{The number of times to rotate the image.}

\item{index}{A character value specifying the index for conversion to binary image.}

\item{lower_size}{Plant images often contain dirt and dust. To prevent dust from
affecting the image analysis, objects with lesser than 10\% of the mean of all objects
are removed. Set \code{lower_limit = 0} to keep all the objects.}

\item{watershed}{If \code{TRUE} (default) performs watershed-based object detection.}

\item{invert}{Inverts the binary image if desired.}

\item{fill_hull}{Fill holes in the binary image? Defaults to \code{FALSE}.}

\item{opening, closing, filter, erode, dilate}{\strong{Morphological operations (brush
size)}, applied in hierarchical order.}

\item{threshold}{The threshold method (\code{"Otsu"}, \code{"Adaptive"}, or numeric).}

\item{extension}{Neighborhood radius for watershed segmentation.}

\item{tolerance}{Minimum height for watershed segmentation.}

\item{object_size}{The size of the object (\code{"small"}, \code{"medium"}, \code{"large"}, \code{"elarge"}).
Used to automatically set up \code{tolerance} and \code{extension} parameters.}

\item{edge}{The number of pixels to be added in the edge of the segmented
object. Defaults to 5.}

\item{remove_bg}{If \code{TRUE}, the pixels that are not part of objects are
converted to white.}

\item{parallel}{If \code{TRUE} processes the images asynchronously in parallel.}

\item{workers}{Number of parallel processes.}

\item{verbose}{If \code{TRUE} (default) shows progress summary in console.}
}
\value{
A \code{NULL} object.
}
\description{
Givin an image with multiple objects, \code{object_export()} will split the
objects into a list of objects using \code{\link[=object_split]{object_split()}} and then export them to
multiple images into the current working directory (or a subfolder). Batch
processing is performed by declaring a file name pattern that matches the
images within the working directory.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("potato_leaves.jpg")
object_export(img,
              remove_bg = TRUE)
}
}
