% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shp.R
\name{object_split_shp}
\alias{object_split_shp}
\title{Splits image objects based on a shape file}
\usage{
object_split_shp(img, rows = 1, cols = 1, ...)
}
\arguments{
\item{img}{An object of class \code{Image}}

\item{rows}{The number of desired rows in the grid. Defaults to \code{1}.}

\item{cols}{The number of desired columns in the grid. Defaults to \code{1}.}

\item{...}{Other arguments passed on to \code{\link[=image_shp]{image_shp()}}}
}
\value{
A list of \code{Image} objects
}
\description{
Here, \code{\link[=image_shp]{image_shp()}} is used to create a shape file based on the desired
number of rows and columns. Then, using the object coordinates, a list of
\code{Image} objects is created.
}
\examples{
if(interactive()){
library(pliman)
flax <- image_pliman("flax_leaves.jpg", plot = TRUE)
objects <- object_split_shp(flax, rows = 3, cols = 5)
image_combine(objects)
}
}
