% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shp.R
\name{object_split_shp}
\alias{object_split_shp}
\title{Splits image objects based on a shape file}
\usage{
object_split_shp(
  img,
  nrow = 1,
  ncol = 1,
  interactive = FALSE,
  only_shp = FALSE,
  ...
)
}
\arguments{
\item{img}{An object of class \code{Image}}

\item{nrow}{The number of desired rows in the grid. Defaults to \code{1}.}

\item{ncol}{The number of desired columns in the grid. Defaults to \code{1}.}

\item{interactive}{If \code{FALSE} (default) the grid is created automatically
based on the image dimension and number of rows/columns. If \code{interactive = TRUE}, users must draw points at the diagonal of the desired bounding box
that will contain the grid.}

\item{only_shp}{If \code{TRUE} returns only the shapefiles with the coordinates
for each image. If \code{FALSE} (default) returns the splitted image according
to \code{nrow} and \code{ncol} arguments.}

\item{...}{Other arguments passed on to \code{\link[=image_shp]{image_shp()}}}
}
\value{
A list of \code{Image} objects
}
\description{
Here, \code{\link[=image_shp]{image_shp()}} is used to create a shape file based on the desired
number of rows and columns. Then, using the object coordinates, a list of
\code{Image} objects is created.
}
\examples{
if(interactive()){
library(pliman)
flax <- image_pliman("flax_leaves.jpg", plot = TRUE)
objects <- object_split_shp(flax, nrow = 3, ncol = 5)
image_combine(objects$imgs)
}
}
