% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{palettes}
\alias{palettes}
\alias{image_palette}
\title{Create image palettes}
\usage{
image_palette(
  img,
  pattern = NULL,
  npal = 5,
  proportional = TRUE,
  colorspace = c("rgb", "hsb"),
  remove_bg = FALSE,
  index = "B",
  filter_order = c("erode", "dilate", "opening", "closing", "filter", "fill_hull"),
  plot = TRUE,
  save_image = FALSE,
  prefix = "proc_",
  dir_original = NULL,
  dir_processed = NULL,
  return_pal = FALSE,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{img}{An image object.}

\item{pattern}{A pattern of file name used to identify images to be imported.}

\item{npal}{The number of color palettes.}

\item{proportional}{Creates a joint palette with proportional size equal to
the number of pixels in the image? Defaults to \code{TRUE}.}

\item{colorspace}{The color space to produce the clusters. Defaults to \code{rgb}.
If \code{hsb}, the color space is first converted from RGB > HSB before k-means
algorithm be applied.}

\item{remove_bg}{Remove background from the color palette? Defaults to
\code{FALSE}.}

\item{index}{An image index used to remove the background, passed to
\code{\link[=image_binary]{image_binary()}}.}

\item{filter_order}{Hierarchical order of morphological operations.}

\item{plot}{Plot the generated palette? Defaults to \code{TRUE}.}

\item{save_image}{Save the image after processing?}

\item{prefix}{Prefix for processed images.}

\item{dir_original, dir_processed}{Directories for original and processed images.}

\item{return_pal}{Return the color palette image? Defaults to \code{FALSE}.}

\item{parallel}{If TRUE processes the images asynchronously (in parallel) in
separate R sessions running in the background on the same machine.}

\item{workers}{Number of parallel processes.}

\item{verbose}{If \code{TRUE} (default) shows progress summary in console.}
}
\value{
\code{image_palette()} returns a list with two elements:
\itemize{
\item \code{palette_list} A list with \code{npal} color palettes of class \code{image}.
\item \code{joint} An object of class \code{image} with the color palettes
\item \code{proportions} The proportion of the entire image corresponding to each color in the palette
\item \code{rgbs} The average RGB value for each palette
}
}
\description{
\code{image_palette()}  creates image palettes by applying the k-means algorithm
to the RGB values.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("sev_leaf.jpg")
pal <- image_palette(img, npal = 5)
}


}
