% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pliman_images}
\alias{pliman_images}
\title{Sample images}
\format{
\verb{*.jpg} format
\itemize{
\item \code{la_back.jpg} A cyan palette representing the background of images
la_pattern, la_leaves, and soybean_touch.
\item \code{la_leaf.jpg} A sample of the leaves in \code{la_leaves}
\item \code{la_leaves.JPG} Tree leaves with a sample of known area.
\item \code{la_pattern.JPG} Tree leaves with a yellow sample of known area.
\item \code{la_temp.jpg} The yellow sample of \code{la_pattern.JPG}.
\item \code{objects_300dpi.jpg} An image with 300 dpi resolution.
\item \code{sev_leaf.jpg} A soybean leaf with a blue background.
\item \code{sev_leaf_nb.jpg} A soybean leaf without background.
\item \code{sev_back.jpg} A blue palette representing the background of \code{sev_leaf}.
\item \code{sev_healthy.jpg} Healthy area of \code{sev_leaf}.
\item \code{sev_sympt.jpg} The symptomatic area \code{sev_leaf}.
\item \code{soy_green.jpg} Soybean grains with a white background.
\item \code{soybean_grain.jpg} A sample palette of the grains in \code{soy_green}.
\item \code{soybean_touch.jpg} Soybean grains with a cyan background touching one each other.
}
}
\source{
Personal data
}
\description{
Sample images installed with the \pkg{pliman} package
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{images}
