% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{prepare_to_shp}
\alias{prepare_to_shp}
\title{Prepare images to analyze_objects_shp()}
\usage{
prepare_to_shp(img, align = "vertical")
}
\arguments{
\item{img}{A \code{Image} object}

\item{align}{The desired alignment. Either \code{"vertical"} (default) or
\code{"horizontal"}.}
}
\value{
An aligned and cropped \code{Image} object.
}
\description{
It is a simple wrapper around \code{\link[=image_align]{image_align()}} and \code{\link[=image_crop]{image_crop()}}. In this case, only the option \code{viewer = "base"} is used. To use \code{viewer = "mapview"}, please, use such functions separately.
}
\examples{
if(interactive()){
img <- image_pliman("flax_leaves.jpg")
prepare_to_shp(img)
}
}
