% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mosaic.R
\name{shapefile_plot}
\alias{shapefile_plot}
\title{A wrapper around terra::plot()}
\usage{
shapefile_plot(shapefile, ...)
}
\arguments{
\item{shapefile}{An SpatVector of sf object.}

\item{...}{Further arguments passed on to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
A \code{NULL} object
}
\description{
Plot the values of a SpatVector
}
\examples{
library(pliman)
r <- shapefile_input(system.file("ex/lux.shp", package="terra"))
shapefile_plot(r)
}
