% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{utils_image}
\alias{utils_image}
\alias{image_import}
\alias{image_export}
\alias{image_show}
\title{Import, export and display images}
\usage{
image_import(image, ...)

image_export(image, name, ...)

image_show(image)
}
\arguments{
\item{image}{\itemize{
\item For \code{image_import()}, a character vector of file names or URLs.
\item For \code{image_export()}, an Image object or an array.
}}

\item{...}{Alternative arguments passed to the corresponding functions from
the \code{jpeg}, \code{png}, and \code{tiff} packages.}

\item{name}{An string specifying the name of the image.}
}
\description{
Import images from files and URLs, write images to files and show images.
}
\examples{
library(pliman)
img <- image_import(system.file("tmp_images", "sev3.png", package = "pliman"))
}
