% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R, R/utils_morphological.R
\name{utils_transform}
\alias{utils_transform}
\alias{image_autocrop}
\alias{image_crop}
\alias{image_dimension}
\alias{image_rotate}
\alias{image_horizontal}
\alias{image_vertical}
\alias{image_hreflect}
\alias{image_vreflect}
\alias{image_trim}
\alias{image_skeleton}
\alias{image_thinning}
\alias{image_filter}
\alias{image_blur}
\alias{image_contrast}
\alias{image_erode}
\alias{image_dilate}
\alias{image_fill_hull}
\alias{image_opening}
\alias{image_remove_small}
\alias{image_closing}
\title{Spatial transformations}
\usage{
image_autocrop(
  img,
  index = "NB",
  edge = 5,
  opening = FALSE,
  closing = FALSE,
  filter = FALSE,
  invert = FALSE,
  threshold = "Otsu",
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_crop(
  img,
  width = NULL,
  height = NULL,
  viewer = get_pliman_viewer(),
  downsample = NULL,
  max_pixels = 1e+06,
  show = "rgb",
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_dimension(img, parallel = FALSE, workers = NULL, verbose = TRUE)

image_rotate(
  img,
  angle,
  bg_col = "white",
  bg_color = NULL,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = TRUE
)

image_horizontal(
  img,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_vertical(
  img,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_hreflect(
  img,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_vreflect(
  img,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_trim(
  img,
  edge = NULL,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_skeleton(
  img,
  kern = NULL,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE,
  ...
)

image_thinning(
  img,
  niter = 3,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE,
  ...
)

image_filter(
  img,
  size = 2,
  cache = 512,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_blur(
  img,
  sigma = 3,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_contrast(
  img,
  nx = 8,
  ny = 8,
  clip_limit = 3,
  bins = 256,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_erode(
  img,
  size = NULL,
  shape = "disc",
  external = FALSE,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_dilate(
  img,
  size = NULL,
  shape = "disc",
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_fill_hull(
  img,
  max_size = NULL,
  min_neck_dist = 0,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_opening(
  img,
  size = NULL,
  shape = "disc",
  external = FALSE,
  return_exact = FALSE,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)

image_remove_small(
  img,
  min_size = NULL,
  rel_size = NULL,
  watershed = TRUE,
  plot = FALSE
)

image_closing(
  img,
  size = NULL,
  shape = "disc",
  external = FALSE,
  parallel = FALSE,
  workers = NULL,
  verbose = TRUE,
  plot = FALSE
)
}
\arguments{
\item{img}{An \code{image} object or a list of \code{image} objects. Must be binary
(\code{storage.mode = "logical"}).}

\item{index}{The index to segment the image. See \code{\link[=image_index]{image_index()}} for more
details. Defaults to \code{"NB"} (normalized blue).}

\item{edge}{\itemize{
\item for \code{\link[=image_autocrop]{image_autocrop()}} the number of pixels in the edge of the cropped
image. If \code{edge = 0} the image will be cropped to create a bounding rectangle
(x and y coordinates) around the image objects.
\item for \code{\link[=image_trim]{image_trim()}}, the number of pixels removed from the edges. By
default, 20 pixels are removed from all the edges.
}}

\item{opening, closing, filter}{\strong{Morphological operations (brush size)}
\itemize{
\item \code{opening} performs an erosion followed by a dilation. This helps to
remove small objects while preserving the shape and size of larger objects.
\item \code{closing} performs a dilatation followed by an erosion. This helps to
fill small holes while preserving the shape and size of larger objects.
\item \code{filter} performs median filtering in the binary image. Provide a positive
integer > 1 to indicate the size of the median filtering. Higher values are
more efficient to remove noise in the background but can dramatically impact
the perimeter of objects, mainly for irregular perimeters such as leaves
with serrated edges.
}

Hierarchically, the operations are performed as opening > closing > filter.
The value declared in each argument will define the brush size.}

\item{invert}{Inverts the binary image if desired. This is useful to process
images with a black background. Defaults to \code{FALSE}. If \code{reference = TRUE}
is use, \code{invert} can be declared as a logical vector of length 2 (eg.,
\verb{invert = c(FALSE, TRUE}). In this case, the segmentation of objects and
reference from the foreground using \code{back_fore_index} is performed using the
default (not inverted), and the segmentation of objects from the reference
is performed by inverting the selection (selecting pixels higher than the
threshold).}

\item{threshold}{The theshold method to be used.
\itemize{
\item By default (\code{threshold = "Otsu"}), a threshold value based
on Otsu's method is used to reduce the grayscale image to a binary image. If
a numeric value is informed, this value will be used as a threshold.
\item If \code{threshold = "adaptive"}, adaptive thresholding (Shafait et al. 2008)
is used, and will depend on the \code{k} and \code{windowsize} arguments.
\item If any non-numeric value different than \code{"Otsu"} and \code{"adaptive"} is used,
an iterative section will allow you to choose the threshold based on a
raster plot showing pixel intensity of the index.
}}

\item{parallel}{Logical. If \code{TRUE}, processes the list of images in parallel.}

\item{workers}{Number of parallel workers. If \code{NULL}, defaults to 40\% of
available cores.}

\item{verbose}{Logical. If \code{TRUE}, prints progress messages.}

\item{plot}{Logical. If \code{TRUE}, plots the resulting image(s).}

\item{width, height}{\itemize{
\item For \code{image_resize()} the Width and height of the resized image. These arguments
can be missing. In this case, the image is resized according to the
relative size informed in \code{rel_size}.
\item For \code{image_crop()} a numeric vector indicating the pixel range (x and y,
respectively) that will be maintained in the cropped image, e.g., width =
100:200
}}

\item{viewer}{The viewer option. If not provided, the value is retrieved
using \code{\link[=get_pliman_viewer]{get_pliman_viewer()}}. This option controls the type of viewer to use
for interactive plotting. The available options are "base" and "mapview".
If set to "base", the base R graphics system is used for interactive
plotting. If set to "mapview", the mapview package is used. To set this
argument globally for all functions in the package, you can use the
\code{\link[=set_pliman_viewer]{set_pliman_viewer()}} function. For example, you can run
\code{set_pliman_viewer("mapview")} to set the viewer option to "mapview" for
all functions.}

\item{downsample}{integer; for each dimension the number of
pixels/lines/bands etc that will be skipped; Defaults to \code{NULL}, which will
find the best downsampling factor to approximate the \code{max_pixels} value.}

\item{max_pixels}{integer > 0. Maximum number of cells to use for the plot.
If \code{max_pixels < npixels(img)}, regular sampling is used before plotting.}

\item{show}{How to plot in mapview viewer, either \code{"rgb"} or \code{"index"}.}

\item{angle}{The rotation angle in degrees.}

\item{bg_col, bg_color}{Color used to fill the background pixels, defaults to \code{"white"}.}

\item{top, bottom, left, right}{The number of pixels removed from \code{top},
\code{bottom}, \code{left}, and \code{right} when using \code{\link[=image_trim]{image_trim()}}.}

\item{kern}{An \code{image} object or an array, containing the structuring
element. Defaults to a brushe generated with \code{\link[=make_brush]{make_brush()}}.}

\item{...}{Additional arguments passed on to \code{\link[=image_binary]{image_binary()}}.}

\item{niter}{The number of iterations to perform in the thinning procedure.
Defaults to 3. Set to \code{NULL} to iterate until the binary image is no longer
changing.}

\item{size}{The size of the structuring element (radius in pixels). Default is \code{NULL},
which automatically determines the size based on image dimensions.}

\item{cache}{The the L2 cache size of the system CPU in kB (integer).
Defaults to \code{512}.}

\item{sigma}{A numeric denoting the standard deviation of the Gaussian filter
used for blurring. Defaults to \code{3}.}

\item{nx}{Number of contextual regions in the x-direction (tile grid) for CLAHE contrast enhancement. Defaults to 8.}

\item{ny}{Number of contextual regions in the y-direction (tile grid) for CLAHE contrast enhancement. Defaults to 8.}

\item{clip_limit}{Normalized contrast limit value for CLAHE contrast enhancement. Defaults to 3.}

\item{bins}{Number of histogram bins used for CLAHE contrast enhancement. Defaults to 256.}

\item{shape}{A character string specifying the shape of the structuring element.
Either \code{"disc"} (default) or \code{"square"}.}

\item{external}{Logical. If \code{TRUE}, erosion does not consume pixels inside
internal holes (passed to \code{image_erode()}).}

\item{return_exact}{Logical. If \code{TRUE}, returns the exact original pixels and
boundaries of objects in \code{img} that survived the opening. Uses internal
watershed partitioning to isolate and eliminate noise touching large
objects, preserving 100\% of sharp corners and geometric features of
surviving objects.}

\item{min_size}{Minimum area (in pixels) for an object to be retained.}

\item{rel_size}{Relative threshold as a fraction of the mean object area (e.g. \code{0.1} for 10\% of mean area).}

\item{watershed}{Logical. Default \code{TRUE}. Applies watershed segmentation before measuring component sizes,
ensuring small noise projections touching large objects are split and removed.}
}
\value{
\itemize{
\item \code{image_skeleton()} returns a binary \code{image} object.
\item All other functions returns a  modified version of \code{image} depending on the
\verb{image_*()} function used.
\item If \code{image} is a list, a list of the same length will be returned.
}

A binary \code{image} object or a list of \code{image} objects.

A binary \code{image} object or a list of \code{image} objects.

A binary \code{image} object or a list of \code{image} objects.

A binary \code{image} object or a list of \code{image} objects.

A binary \code{image} object or a list of binary \code{image} objects.

A binary \code{image} object or a list of \code{image} objects.
}
\description{
Performs image rotation and reflection
\itemize{
\item \verb{image autocrop()} Crops automatically  an image to the area of objects.
\item \code{image_crop()} Crops an image to the desired area.
\item \code{image_trim()} Remove pixels from the edges of an image (20 by default).
\item \code{image_dimension()} Gives the dimension (width and height) of an image.
\item \code{image_rotate()} Rotates the image clockwise by the given angle.
\item \code{image_horizontal()} Converts (if needed) an image to a horizontal image.
\item \code{image_vertical()} Converts (if needed) an image to a vertical image.
\item \code{image_hreflect()} Performs horizontal reflection of the \code{image}.
\item \code{image_vreflect()} Performs vertical reflection of the \code{image}.
\item \code{image_resize()} Resize the \code{image}. See more at \code{\link[=image_resize]{image_resize()}}.
\item \code{image_contrast()} Improve contrast locally by performing adaptive
histogram equalization.
\item \code{image_dilate()} Performs image dilatation.
\item \code{image_erode()} Performs image erosion.
\item \code{image_opening()} Performs an erosion followed by a dilation.
\item \code{image_closing()} Performs a dilation followed by an erosion.
\item \code{image_filter()} Performs median filtering in constant time.
\item \code{image_blur()} Performs blurring filter of images.
\item \code{image_skeleton()} Performs image skeletonization.
}

Performs morphological erosion of a binary image.

Performs morphological dilation of a binary image.

Fills internal holes in a binary image.

Performs morphological opening of a binary image (erosion followed by dilation).
When \code{return_exact = TRUE}, morphological reconstruction via watershed-partitioned seed matching is performed.
This isolates noise touching large objects into separate watershed labels, identifies surviving labels using the un-dilated eroded marker,
and restores 100\% of the exact original boundaries and shapes of surviving objects without rounding corners or eroding tips.

Removes small noise objects from a binary image without altering or eroding
the boundaries of larger objects. Filtering can be based on an absolute pixel count (\code{min_size})
or a relative fraction of the mean object area (\code{rel_size}, e.g. \code{0.1} for objects smaller than 10\% of average object area).

Performs morphological closing of a binary image. Closing is defined as a
dilation followed by an erosion. It is useful for filling small holes or
connecting nearby objects.
}
\examples{
if (interactive()) {
library(pliman)
img <- image_pliman("sev_leaf.jpg")
plot(img)
img <- image_resize(img, 50)
img1 <- image_rotate(img, 45)
img2 <- image_hreflect(img)
img3 <- image_vreflect(img)
img4 <- image_vertical(img)
image_combine(img1, img2, img3, img4)
}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
